refactor: move conversions to the p2p crate
