+++ crates/common/src/transaction.rs
@@ -1,3 +1,4 @@
+use fake::Dummy;
 use pathfinder_crypto::hash::{HashChain as PedersenHasher, PoseidonHasher};
 use pathfinder_crypto::Felt;
 use primitive_types::H256;
@@ -12,7 +13,7 @@ use crate::{
     Tip,
 };
 
-#[derive(Clone, Debug, PartialEq, Eq)]
+#[derive(Clone, Debug, PartialEq, Eq, Dummy)]
 pub struct Transaction {
     pub hash: TransactionHash,
     pub variant: TransactionVariant,
@@ -43,7 +44,7 @@ impl Transaction {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq)]
+#[derive(Clone, Debug, PartialEq, Eq, Dummy)]
 pub enum TransactionVariant {
     DeclareV0(DeclareTransactionV0V1),
     DeclareV1(DeclareTransactionV0V1),
@@ -197,7 +198,7 @@ impl From<L1HandlerTransaction> for TransactionVariant {
     }
 }
 
-#[derive(Clone, Default, Debug, PartialEq, Eq)]
+#[derive(Clone, Default, Debug, PartialEq, Eq, Dummy)]
 pub struct DeclareTransactionV0V1 {
     pub class_hash: ClassHash,
     pub max_fee: Fee,
@@ -206,7 +207,7 @@ pub struct DeclareTransactionV0V1 {
     pub sender_address: ContractAddress,
 }
 
-#[derive(Clone, Default, Debug, PartialEq, Eq)]
+#[derive(Clone, Default, Debug, PartialEq, Eq, Dummy)]
 pub struct DeclareTransactionV2 {
     pub class_hash: ClassHash,
     pub max_fee: Fee,
@@ -216,7 +217,7 @@ pub struct DeclareTransactionV2 {
     pub compiled_class_hash: CasmHash,
 }
 
-#[derive(Clone, Default, Debug, PartialEq, Eq)]
+#[derive(Clone, Default, Debug, PartialEq, Eq, Dummy)]
 pub struct DeclareTransactionV3 {
     pub class_hash: ClassHash,
     pub nonce: TransactionNonce,
@@ -231,7 +232,7 @@ pub struct DeclareTransactionV3 {
     pub compiled_class_hash: CasmHash,
 }
 
-#[derive(Clone, Default, Debug, PartialEq, Eq)]
+#[derive(Clone, Default, Debug, PartialEq, Eq, Dummy)]
 pub struct DeployTransactionV0 {
     pub class_hash: ClassHash,
     pub contract_address: ContractAddress,
@@ -239,7 +240,7 @@ pub struct DeployTransactionV0 {
     pub constructor_calldata: Vec<ConstructorParam>,
 }
 
-#[derive(Clone, Default, Debug, PartialEq, Eq)]
+#[derive(Clone, Default, Debug, PartialEq, Eq, Dummy)]
 pub struct DeployTransactionV1 {
     pub class_hash: ClassHash,
     pub contract_address: ContractAddress,
@@ -247,7 +248,7 @@ pub struct DeployTransactionV1 {
     pub constructor_calldata: Vec<ConstructorParam>,
 }
 
-#[derive(Clone, Default, Debug, PartialEq, Eq)]
+#[derive(Clone, Default, Debug, PartialEq, Eq, Dummy)]
 pub struct DeployAccountTransactionV1 {
     pub contract_address: ContractAddress,
     pub max_fee: Fee,
@@ -258,7 +259,7 @@ pub struct DeployAccountTransactionV1 {
     pub class_hash: ClassHash,
 }
 
-#[derive(Clone, Default, Debug, PartialEq, Eq)]
+#[derive(Clone, Default, Debug, PartialEq, Eq, Dummy)]
 pub struct DeployAccountTransactionV3 {
     pub contract_address: ContractAddress,
     pub signature: Vec<TransactionSignatureElem>,
@@ -273,7 +274,7 @@ pub struct DeployAccountTransactionV3 {
     pub class_hash: ClassHash,
 }
 
-#[derive(Clone, Default, Debug, PartialEq, Eq)]
+#[derive(Clone, Default, Debug, PartialEq, Eq, Dummy)]
 pub struct InvokeTransactionV0 {
     pub calldata: Vec<CallParam>,
     pub sender_address: ContractAddress,
@@ -283,7 +284,7 @@ pub struct InvokeTransactionV0 {
     pub signature: Vec<TransactionSignatureElem>,
 }
 
-#[derive(Clone, Default, Debug, PartialEq, Eq)]
+#[derive(Clone, Default, Debug, PartialEq, Eq, Dummy)]
 pub struct InvokeTransactionV1 {
     pub calldata: Vec<CallParam>,
     pub sender_address: ContractAddress,
@@ -292,7 +293,7 @@ pub struct InvokeTransactionV1 {
     pub nonce: TransactionNonce,
 }
 
-#[derive(Clone, Default, Debug, PartialEq, Eq)]
+#[derive(Clone, Default, Debug, PartialEq, Eq, Dummy)]
 pub struct InvokeTransactionV3 {
     pub signature: Vec<TransactionSignatureElem>,
     pub nonce: TransactionNonce,
@@ -306,7 +307,7 @@ pub struct InvokeTransactionV3 {
     pub sender_address: ContractAddress,
 }
 
-#[derive(Clone, Default, Debug, PartialEq, Eq)]
+#[derive(Clone, Default, Debug, PartialEq, Eq, Dummy)]
 pub struct L1HandlerTransaction {
     pub contract_address: ContractAddress,
     pub entry_point_selector: EntryPoint,
@@ -314,25 +315,25 @@ pub struct L1HandlerTransaction {
     pub calldata: Vec<CallParam>,
 }
 
-#[derive(Copy, Clone, Debug, PartialEq, Eq)]
+#[derive(Copy, Clone, Debug, PartialEq, Eq, Dummy)]
 pub enum EntryPointType {
     External,
     L1Handler,
 }
 
-#[derive(Copy, Clone, Debug, Default, PartialEq, Eq)]
+#[derive(Copy, Clone, Debug, Default, PartialEq, Eq, Dummy)]
 pub struct ResourceBounds {
     pub l1_gas: ResourceBound,
     pub l2_gas: ResourceBound,
 }
 
-#[derive(Copy, Clone, Debug, Default, PartialEq, Eq)]
+#[derive(Copy, Clone, Debug, Default, PartialEq, Eq, Dummy)]
 pub struct ResourceBound {
     pub max_amount: ResourceAmount,
     pub max_price_per_unit: ResourcePricePerUnit,
 }
 
-#[derive(Copy, Clone, Debug, Default, PartialEq, Eq)]
+#[derive(Copy, Clone, Debug, Default, PartialEq, Eq, Dummy)]
 pub enum DataAvailabilityMode {
     #[default]
     L1,
